bots/requirements.txt
pillow
atproto
requests
python-dotenv